chore: split test files
